requirements.txt
streamlit==1.31.0
tensorflow==2.15.0
keras==3.0.5
numpy==1.24.3
opencv-python-headless==4.9.0.80
Pillow==10.2.0
matplotlib==3.8.3